Develop

+++ src/main/java/com/actilazion/ariesreportingproject/dto/response/AdminOverviewResponse.java
@@ -2,6 +2,7 @@ package com.actilazion.ariesreportingproject.dto.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -9,7 +10,7 @@ public record AdminOverviewResponse(
         String periodFrom,
         String periodTo,
         long totalTransactions,
-        double totalVolume,
+        BigDecimal totalVolume,
         long completedCount,
         long failedCount,
         long pendingCount,
@@ -19,7 +20,7 @@ public record AdminOverviewResponse(
     public record DailyVolume(
             String day,
             long txCount,
-            double volume
+            BigDecimal volume
     ) {
     }
 }
